To Master - Get Olympian Stats

+++ db/migrations/20190128225204_initial.js
@@ -5,9 +5,9 @@ exports.up = function(knex) {
       table.increments('id').primary();
       table.string('name');
       table.string('sex');
-      table.integer('age');
-      table.integer('height');
-      table.integer('weight');
+      table.smallint('age');
+      table.smallint('height');
+      table.smallint('weight');
       table.string('team');
       table.string('games');
       table.string('sport');
